fix: split organization-utils to separate server-only code from client-safe code
The build was failing because organization-utils.ts imported serverAuthClient
from server-client.ts, which imports next/headers. This created a dependency
chain that pulled next/headers into client components.

Split into:
- organization-utils.ts: client-safe functions only
- organization-utils.server.ts: server-only functions using serverAuthClient
